+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🎉 feat(slices-sorted_last_index_by): sorted_last_index_by function added
 * ✅ test(slices-sorted_last_index): sorted_last_index test cases and benchmarks added
 * 🎉 feat(slices-sorted_last_index): sorted_last_index function added
 * 🐛 fix(slices): sorted_index* functions meant to return lowest index of passed value but it was returning the highest